+++ firmware/defmt-test/macros/src/lib.rs
@@ -15,6 +15,16 @@ pub fn tests(args: TokenStream, input: TokenStream) -> TokenStream {
     }
 }
 
+macro_rules! fn_state_signature_msg {
+    ($name:literal) => {
+        concat!(
+            "`#[",
+            $name,
+            "]` function must have signature `fn()` or `fn(state: &mut T)`"
+        )
+    };
+}
+
 fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStream> {
     if !args.is_empty() {
         return Err(parse::Error::new(
@@ -35,6 +45,7 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
     };
 
     let mut init = None;
+    let mut teardown = None;
     let mut before_each = None;
     let mut after_each = None;
     let mut tests = vec![];
@@ -49,6 +60,9 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
                 f.attrs.retain(|attr| {
                     if attr.path().is_ident("init") {
                         test_kind = Some(Attr::Init);
+                        false
+                    } else if attr.path().is_ident("teardown") {
+                        test_kind = Some(Attr::Teardown);
                         false
                     } else if attr.path().is_ident("test") {
                         test_kind = Some(Attr::Test);
@@ -75,7 +89,7 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
                     None => {
                         return Err(parse::Error::new(
                             f.span(),
-                            "function requires `#[init]`, `#[before_each]`, `#[after_each]`, or `#[test]` attribute",
+                            "function requires `#[init]`, `#[teardown]`, `#[before_each]`, `#[after_each]`, or `#[test]` attribute",
                         ));
                     }
                 };
@@ -106,7 +120,7 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
                         if check_fn_sig(&f.sig).is_err() || !f.sig.inputs.is_empty() {
                             return Err(parse::Error::new(
                                 f.sig.ident.span(),
-                                "`#[init]` function must have signature `fn() [-> Type]` (the return type is optional)",
+                                "`#[init]` function must have signature `fn()` or `fn() -> T`",
                             ));
                         }
 
@@ -118,11 +132,61 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
                         init = Some(Init { func: f, state });
                     }
 
+                    Attr::Teardown => {
+                        if teardown.is_some() {
+                            return Err(parse::Error::new(
+                                f.sig.ident.span(),
+                                "only a single `#[teardown]` function can be defined",
+                            ));
+                        }
+
+                        if should_error {
+                            return Err(parse::Error::new(
+                                f.sig.ident.span(),
+                                "`#[should_error]` is not allowed on the `#[teardown]` function",
+                            ));
+                        }
+
+                        if ignore {
+                            return Err(parse::Error::new(
+                                f.sig.ident.span(),
+                                "`#[ignore]` is not allowed on the `#[teardown]` function",
+                            ));
+                        }
+
+                        if check_fn_sig(&f.sig).is_err() || f.sig.inputs.len() > 1 {
+                            return Err(parse::Error::new(
+                                f.sig.ident.span(),
+                                fn_state_signature_msg!("teardown"),
+                            ));
+                        }
+
+                        let input = if f.sig.inputs.len() == 1 {
+                            let arg = &f.sig.inputs[0];
+
+                            // NOTE we cannot check the argument type matches `init.state` at this
+                            // point
+                            if let Some(ty) = get_mutable_reference_type(arg).cloned() {
+                                Some(Input { ty })
+                            } else {
+                                // was not `&mut T`
+                                return Err(parse::Error::new(
+                                    arg.span(),
+                                    "parameter must be a mutable reference (`&mut $Type`)",
+                                ));
+                            }
+                        } else {
+                            None
+                        };
+
+                        teardown = Some(Teardown { func: f, input });
+                    }
+
                     Attr::Test => {
                         if check_fn_sig(&f.sig).is_err() || f.sig.inputs.len() > 1 {
                             return Err(parse::Error::new(
                                 f.sig.ident.span(),
-                                "`#[test]` function must have signature `fn(state: &mut Type)` (parameter is optional)",
+                                fn_state_signature_msg!("test"),
                             ));
                         }
 
@@ -177,7 +241,7 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
                         if check_fn_sig(&f.sig).is_err() || f.sig.inputs.len() > 1 {
                             return Err(parse::Error::new(
                                 f.sig.ident.span(),
-                                "`#[before_each]` function must have signature `fn(state: &mut Type)` (parameter is optional)",
+                                fn_state_signature_msg!("before_each"),
                             ));
                         }
 
@@ -226,7 +290,7 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
                         if check_fn_sig(&f.sig).is_err() || f.sig.inputs.len() > 1 {
                             return Err(parse::Error::new(
                                 f.sig.ident.span(),
-                                "`#[after_each]` function must have signature `fn(state: &mut Type)` (parameter is optional)",
+                                fn_state_signature_msg!("after_each"),
                             ));
                         }
 
@@ -271,6 +335,39 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
             Some(quote!(#init_func)),
             Some(quote!(#[allow(dead_code)] let mut state = #init_ident();)),
         )
+    } else {
+        (None, None)
+    };
+
+    let (teardown_fn, teardown_expr) = if let Some(teardown) = teardown {
+        let teardown_func = &teardown.func;
+        let teardown_ident = &teardown.func.sig.ident;
+        let span = teardown.func.sig.ident.span();
+
+        let call = if let Some(input) = teardown.input.as_ref() {
+            if let Some(state) = &state_ty {
+                if input.ty != **state {
+                    return Err(parse::Error::new(
+                        input.ty.span(),
+                        format!(
+                            "this type must match `#[init]`s return type: {}",
+                            type_ident(state)
+                        ),
+                    ));
+                }
+            } else {
+                return Err(parse::Error::new(
+                    span,
+                    "no state was initialized by `#[init]`; signature must be `fn()`",
+                ));
+            }
+
+            quote!(#teardown_ident(&mut state))
+        } else {
+            quote!(#teardown_ident())
+        };
+
+        (Some(quote!(#teardown_func)), Some(quote!(#call;)))
     } else {
         (None, None)
     };
@@ -433,10 +530,15 @@ fn tests_impl(args: TokenStream, input: TokenStream) -> parse::Result<TokenStrea
             )*
 
             defmt::println!("all tests passed!");
+
+            #teardown_expr
+
             #krate::export::exit()
         }
 
         #init_fn
+
+        #teardown_fn
 
         #before_each_fn
 
@@ -454,6 +556,7 @@ enum Attr {
     AfterEach,
     BeforeEach,
     Init,
+    Teardown,
     Test,
 }
 
@@ -470,6 +573,11 @@ struct BeforeEach {
 struct Init {
     func: ItemFn,
     state: Option<Box<Type>>,
+}
+
+struct Teardown {
+    func: ItemFn,
+    input: Option<Input>,
 }
 
 struct Test {