+++ firmware/defmt-test/macros/tests/ui/teardown-has-immutable-param.rs
@@ -1,0 +1,9 @@
+fn main() {}
+
+#[defmt_test_macros::tests]
+mod tests {
+    #[teardown]
+    fn say(name: &str) {
+        assert_eq!("name", name);
+    }
+}