+++ firmware/defmt-test/macros/tests/ui/teardown-has-incompatible-init-signature.rs
@@ -1,0 +1,14 @@
+fn main() {}
+
+#[defmt_test_macros::tests]
+mod tests {
+    #[init]
+    fn init() -> u32 {
+        0_u32
+    }
+
+    #[teardown]
+    fn say(value: &mut u16) {
+        assert!(true);
+    }
+}