+++ firmware/defmt-test/macros/tests/ui/teardown-has-non-empty-signature.rs
@@ -1,0 +1,14 @@
+fn main() {}
+
+#[defmt_test_macros::tests]
+mod tests {
+    #[init]
+    fn init() {
+        // empty
+    }
+
+    #[teardown]
+    fn teardown(arg: &mut u8) {
+        assert!(true);
+    }
+}